Preliminary work to try native pdf rendering, serializing displaylist from JNI to flatbuffers, and parsing them in java to render them to a canvas. The flatbuffer schema is mostly complete, except for mesh shades, but it seems a bit too short to be true. I'll then need to serialize fz_displaylist using flatbuffers and render that in Java.
If that don't work, I'll render the pdf using bitmap tiles computed using the displaylist on a tileview (I'll need it either way to display huge comics image for example, when they are bigger than Canvas.get).
Cleaned PDF API a little, and removed link and outline stuff as we'll pass that from JNI to Java using flatbuffers, but later.